refactor(pipeline): extract dispatch orchestrator + world-change parsing from routes/games.py
Move _dispatch_pipeline (complexity 39, 272 lines) and its exclusive
background-task helpers (_generate_title_bg, _write_snapshot_bg,
_regen_summary_bg) from routes/games.py into pipeline/orchestrator.py
as public dispatch_pipeline.

Extract world-change parsing helpers (attribute-type mapping,
translate_world_updates, parse_relationship_delta) into new
pipeline/world_changes.py module.

Reduces routes/games.py from 2,191 to 1,609 lines (-27%).
All ruff/pyright clean. Call sites updated to use new public names.
