+++ src/tta/api/routes/games.py
@@ -30,12 +30,10 @@ from tta.api.errors import AppError
 from tta.api.sse import SseEventBuffer
 from tta.config import Settings, get_settings
 from tta.errors import ErrorCategory
-from tta.logging import bind_context
 from tta.models.game import (
     CreateGameRequest,
     DeleteGameRequest,
     GameData,
-    GameState,
     GameStatus,
     GameSummary,
     PaginationMeta,
@@ -47,9 +45,6 @@ from tta.models.game import (
 from tta.models.player import Player
 from tta.models.turn import TurnState, TurnStatus
 from tta.models.world import (
-    RelationshipChange,
-    WorldChange,
-    WorldChangeType,
     WorldSeed,
 )
 from tta.observability.metrics import (
@@ -62,510 +57,20 @@ from tta.observability.metrics import (
     TURN_STORAGE_OPS_DURATION,
 )
 from tta.persistence.redis_session import set_active_session
-from tta.pipeline.orchestrator import run_pipeline
-from tta.privacy.cost import get_cost_tracker, reset_cost_tracker
+from tta.pipeline.orchestrator import (
+    _regen_summary_bg,
+    dispatch_pipeline,
+)
+from tta.pipeline.world_changes import (
+    translate_world_updates,
+)
 from tta.transport import SSETransport
 
 log = structlog.get_logger()
 
 router = APIRouter(prefix="/games", tags=["games"])
 
-# Mapping from LLM-extracted attribute keywords to WorldChangeType
-_ATTRIBUTE_TYPE_MAP: dict[str, WorldChangeType] = {
-    "location": WorldChangeType.PLAYER_MOVED,
-    "position": WorldChangeType.PLAYER_MOVED,
-    "moved": WorldChangeType.PLAYER_MOVED,
-    "disposition": WorldChangeType.NPC_DISPOSITION_CHANGED,
-    "mood": WorldChangeType.NPC_DISPOSITION_CHANGED,
-    "attitude": WorldChangeType.NPC_DISPOSITION_CHANGED,
-    "state": WorldChangeType.LOCATION_STATE_CHANGED,
-    "status": WorldChangeType.LOCATION_STATE_CHANGED,
-    "locked": WorldChangeType.CONNECTION_LOCKED,
-    "unlocked": WorldChangeType.CONNECTION_UNLOCKED,
-    "taken": WorldChangeType.ITEM_TAKEN,
-    "picked": WorldChangeType.ITEM_TAKEN,
-    "dropped": WorldChangeType.ITEM_DROPPED,
-    "quest_status": WorldChangeType.QUEST_STATUS_CHANGED,
-    "quest": WorldChangeType.QUEST_STATUS_CHANGED,
-    "visibility": WorldChangeType.ITEM_VISIBILITY_CHANGED,
-    "relationship": WorldChangeType.RELATIONSHIP_CHANGED,
-}
-
-
-def _build_typed_payload(change_type: WorldChangeType, item: dict) -> dict:
-    """Build a payload dict with the keys required by validate_change()."""
-    base: dict = {
-        "attribute": str(item.get("attribute") or ""),
-        "old_value": item.get("old_value"),
-        "new_value": item.get("new_value"),
-        "reason": item.get("reason", ""),
-    }
-    # Merge any extra keys the LLM may have provided (e.g. from_id, to_id)
-    for k, v in item.items():
-        if k not in ("entity", "attribute", "old_value", "new_value", "reason"):
-            base[k] = v
-
-    nv = item.get("new_value")
-    ct = change_type
-    if ct == WorldChangeType.PLAYER_MOVED:
-        base.setdefault("from_id", item.get("old_value", ""))
-        base.setdefault("to_id", item.get("new_value", ""))
-    elif ct == WorldChangeType.NPC_DISPOSITION_CHANGED:
-        base.setdefault("disposition", nv if nv is not None else "")
-    elif ct == WorldChangeType.NPC_STATE_CHANGED:
-        base.setdefault("state", nv if nv is not None else "")
-    elif ct == WorldChangeType.NPC_MOVED:
-        base.setdefault("to_location_id", nv if nv is not None else "")
-    elif ct == WorldChangeType.CONNECTION_LOCKED:
-        base.setdefault("from_id", item.get("old_value", ""))
-        base.setdefault("to_id", str(item.get("entity", "")))
-    elif ct == WorldChangeType.CONNECTION_UNLOCKED:
-        base.setdefault("from_id", item.get("old_value", ""))
-        base.setdefault("to_id", str(item.get("entity", "")))
-    elif ct == WorldChangeType.QUEST_STATUS_CHANGED:
-        base.setdefault("new_status", nv if nv is not None else "")
-    elif ct == WorldChangeType.ITEM_VISIBILITY_CHANGED:
-        hidden = nv if isinstance(nv, bool) else str(nv).lower() == "true"
-        base.setdefault("hidden", hidden)
-    elif ct == WorldChangeType.RELATIONSHIP_CHANGED:
-        base.setdefault("dimension", str(item.get("attribute") or "trust"))
-        base.setdefault("direction", str(nv) if nv is not None else "positive")
-    return base
-
-
-def _translate_world_updates(raw: list[dict]) -> list[WorldChange]:
-    """Convert LLM-extracted dicts to WorldChange objects (best-effort)."""
-    changes: list[WorldChange] = []
-    for item in raw:
-        entity = item.get("entity", "")
-        attribute = str(item.get("attribute") or "")
-        if not entity:
-            continue
-        # Infer change type from attribute keywords
-        change_type = WorldChangeType.LOCATION_STATE_CHANGED
-        attr_lower = attribute.lower()
-        for keyword, ct in sorted(
-            _ATTRIBUTE_TYPE_MAP.items(), key=lambda x: -len(x[0])
-        ):
-            if keyword in attr_lower:
-                change_type = ct
-                break
-        changes.append(
-            WorldChange(
-                type=change_type,
-                entity_id=str(entity),
-                payload=_build_typed_payload(change_type, item),
-            )
-        )
-    return changes
-
-
-# Keywords for inferring relationship dimension and direction from LLM output
-_POSITIVE_KEYWORDS = frozenset(
-    {"increase", "improve", "gain", "grow", "positive", "warm", "better", "higher"}
-)
-_NEGATIVE_KEYWORDS = frozenset(
-    {"decrease", "lose", "drop", "worsen", "negative", "cold", "worse", "lower"}
-)
-
-
-def _parse_relationship_delta(payload: dict) -> RelationshipChange:
-    """Convert LLM-extracted relationship payload to a RelationshipChange.
-
-    The payload contains ``dimension`` (attribute like "trust", "fear") and
-    ``direction`` (a descriptive string like "increased", "grew warmer").
-    We map these to a small numeric delta on the appropriate axis.
-    """
-    dimension = payload.get("dimension", "trust").lower()
-    direction_raw = str(payload.get("direction", "positive")).lower()
-
-    # Determine sign: positive or negative shift
-    sign = 1
-    if any(kw in direction_raw for kw in _NEGATIVE_KEYWORDS):
-        sign = -1
-
-    delta = 5 * sign  # modest default step
-
-    trust = 0
-    affinity = 0
-    respect = 0
-    fear = 0
-    familiarity = 3  # any interaction increases familiarity
-
-    if "trust" in dimension:
-        trust = delta
-    elif "affinity" in dimension or "warmth" in dimension:
-        affinity = delta
-    elif "respect" in dimension:
-        respect = delta
-    elif "fear" in dimension:
-        fear = abs(delta) if sign > 0 else -abs(delta)
-    else:
-        # Generic / unmapped → trust + affinity
-        trust = delta
-        affinity = delta
-
-    return RelationshipChange(
-        trust=trust,
-        affinity=affinity,
-        respect=respect,
-        fear=fear,
-        familiarity=familiarity,
-    )
-
-
-async def _dispatch_pipeline(
-    app_state: object,
-    game_id: UUID,
-    turn_id: UUID,
-    turn_number: int,
-    player_input: str,
-    game_state: dict,
-    session_cost_usd: float = 0.0,
-) -> None:
-    """Run the pipeline as a background task and persist results."""
-    deps = app_state.pipeline_deps  # type: ignore[attr-defined]
-    turn_repo = deps.turn_repo
-
-    # Bind game/turn context for correlated logging (S15 §7).
-    bind_context(session_id=game_id, turn_id=turn_id)
-
-    # Seed cost tracker with session total from DB (FR-07.19).
-    reset_cost_tracker(
-        session_id=str(game_id),
-        session_total_usd=session_cost_usd,
-    )
-
-    state = TurnState(
-        session_id=game_id,
-        turn_id=turn_id,
-        turn_number=turn_number,
-        player_input=player_input,
-        game_state=game_state,
-    )
-
-    start = time.monotonic()
-    try:
-        result = await run_pipeline(state, deps)
-    except Exception:
-        log.error("pipeline_dispatch_failed", game_id=str(game_id), exc_info=True)
-        result = state.model_copy(update={"status": TurnStatus.failed})
-
-    elapsed_ms = (time.monotonic() - start) * 1000
-    result = result.model_copy(update={"latency_ms": elapsed_ms})
-
-    # Persist turn result via repository
-    turn_persisted = False
-    try:
-        if (
-            result.status
-            in (
-                TurnStatus.complete,
-                TurnStatus.moderated,
-            )
-            and result.narrative_output
-        ):
-            token_dict = result.token_count.model_dump() if result.token_count else {}
-            await turn_repo.complete_turn(
-                turn_id=turn_id,
-                narrative_output=result.narrative_output,
-                model_used=result.model_used or "unknown",
-                latency_ms=elapsed_ms,
-                token_count=token_dict,
-            )
-            # FR-24.06 item 5: mark moderated turns distinctly
-            if result.status == TurnStatus.moderated:
-                await turn_repo.update_status(turn_id, "moderated")
-            turn_persisted = True
-        else:
-            # FR-23.18: preserve partial narrative on failure
-            await turn_repo.fail_turn(turn_id, narrative_output=result.narrative_output)
-    except Exception:
-        log.error("turn_persist_failed", turn_id=str(turn_id), exc_info=True)
-        # Last-resort: ensure turn exits processing state to prevent
-        # permanent concurrent-turn lock (critique finding #5).
-        try:
-            await turn_repo.update_status(turn_id, "failed")
-        except Exception:
-            log.error(
-                "turn_failsafe_status_update_failed",
-                turn_id=str(turn_id),
-                exc_info=True,
-            )
-
-    # --- FR-27.05/06: Post-turn auto-save (metadata) ---
-    # Gate on turn_persisted to avoid incrementing counters for failed saves.
-    if turn_persisted:
-        # TODO: inject session factory via PipelineDeps instead of
-        # reaching into repo internals (_sf).
-        sf = app_state.pipeline_deps.turn_repo._sf  # type: ignore[attr-defined]
-        try:
-            async with sf() as meta_sess:
-                now = datetime.now(UTC)
-                await meta_sess.execute(
-                    sa.text(
-                        "UPDATE game_sessions "
-                        "SET turn_count = turn_count + 1, "
-                        "last_played_at = :now, updated_at = :now "
-                        "WHERE id = :gid"
-                    ),
-                    {"gid": game_id, "now": now},
-                )
-                await meta_sess.commit()
-        except Exception:
-            log.warning(
-                "auto_save_metadata_failed",
-                game_id=str(game_id),
-                exc_info=True,
-            )
-            # FR-27.07: mark needs_recovery so resume can fix it
-            try:
-                async with sf() as recovery_sess:
-                    await recovery_sess.execute(
-                        sa.text(
-                            "UPDATE game_sessions "
-                            "SET needs_recovery = TRUE "
-                            "WHERE id = :gid"
-                        ),
-                        {"gid": game_id},
-                    )
-                    await recovery_sess.commit()
-            except Exception:
-                log.error(
-                    "needs_recovery_flag_failed",
-                    game_id=str(game_id),
-                    exc_info=True,
-                )
-
-        # FR-27.22: title from opening narrative (turn 1 IS genesis)
-        if turn_number == 1 and result.narrative_output:
-            asyncio.create_task(
-                _generate_title_bg(app_state, game_id, result.narrative_output)
-            )
-
-        # --- FR-07.19: Persist turn cost to session (S07 cost management) ---
-        try:
-            tracker = get_cost_tracker()
-            turn_cost = tracker.turn_cost_usd
-            if turn_cost > 0:
-                async with sf() as cost_sess:
-                    cost_result = await cost_sess.execute(
-                        sa.text(
-                            "UPDATE game_sessions "
-                            "SET total_cost_usd = total_cost_usd + :cost, "
-                            "updated_at = :now "
-                            "WHERE id = :gid "
-                            "RETURNING total_cost_usd, cost_warning_sent"
-                        ),
-                        {
-                            "gid": game_id,
-                            "cost": turn_cost,
-                            "now": datetime.now(UTC),
-                        },
-                    )
-                    cost_row = cost_result.one()
-                    await cost_sess.commit()
-
-                    # Send 80% warning once
-                    settings = get_settings()
-                    cap = settings.session_cost_cap_usd
-                    if (
-                        cap > 0
-                        and not cost_row.cost_warning_sent
-                        and float(cost_row.total_cost_usd)
-                        >= cap * settings.session_cost_warn_pct
-                    ):
-                        async with sf() as warn_sess:
-                            await warn_sess.execute(
-                                sa.text(
-                                    "UPDATE game_sessions "
-                                    "SET cost_warning_sent = true, "
-                                    "updated_at = :now WHERE id = :gid"
-                                ),
-                                {
-                                    "gid": game_id,
-                                    "now": datetime.now(UTC),
-                                },
-                            )
-                            await warn_sess.commit()
-                        log.warning(
-                            "session_cost_warning",
-                            game_id=str(game_id),
-                            total=float(cost_row.total_cost_usd),
-                            cap=cap,
-                        )
-        except Exception:
-            log.warning(
-                "cost_persist_failed",
-                game_id=str(game_id),
-                exc_info=True,
-            )
-
-        # FR-27.20: fire-and-forget summary regen every Nth turn
-        settings = app_state.settings  # type: ignore[attr-defined]
-        if (
-            settings.summary_interval > 0
-            and turn_number % settings.summary_interval == 0
-        ):
-            asyncio.create_task(_regen_summary_bg(app_state, game_id))
-
-        # AC-12.04: fire-and-forget game snapshot every Nth turn
-        if (
-            settings.snapshot_interval > 0
-            and turn_number % settings.snapshot_interval == 0
-        ):
-            asyncio.create_task(
-                _write_snapshot_bg(app_state, game_id, result.game_state, turn_number)
-            )
-
-    # --- Apply world state changes (best-effort) ---
-    if turn_persisted and result.world_state_updates:
-        try:
-            from tta.world.changes import apply_changes
-
-            world_svc = deps.world
-            changes = _translate_world_updates(result.world_state_updates)
-            # Separate relationship changes for dedicated handling (S06 AC-6.4)
-            rel_changes = [
-                c for c in changes if c.type == WorldChangeType.RELATIONSHIP_CHANGED
-            ]
-            other_changes = [
-                c for c in changes if c.type != WorldChangeType.RELATIONSHIP_CHANGED
-            ]
-            if other_changes:
-                await apply_changes(other_changes, world_svc, game_id)
-            # Apply relationship changes via RelationshipService (fire-and-forget)
-            rel_svc = deps.relationship_service
-            if rel_changes and rel_svc is not None:
-                for rc in rel_changes:
-                    try:
-                        delta = _parse_relationship_delta(rc.payload)
-                        await rel_svc.update_relationship(
-                            session_id=str(game_id),
-                            source_id="player",
-                            target_id=rc.entity_id,
-                            change=delta,
-                        )
-                    except Exception:
-                        log.debug(
-                            "relationship_change_skipped",
-                            entity=rc.entity_id,
-                            exc_info=True,
-                        )
-            applied = len(other_changes) + len(rel_changes)
-            if applied:
-                log.info(
-                    "world_changes_applied",
-                    game_id=str(game_id),
-                    count=applied,
-                    relationships=len(rel_changes),
-                )
-        except asyncio.CancelledError:
-            raise
-        except Exception:
-            log.warning(
-                "world_changes_failed_graceful_degradation",
-                game_id=str(game_id),
-                exc_info=True,
-            )
-
-    # Publish result for SSE endpoint
-    try:
-        store = app_state.turn_result_store  # type: ignore[attr-defined]
-        await store.publish(str(turn_id), result)
-    except Exception:
-        log.error(
-            "turn_result_publish_failed",
-            turn_id=str(turn_id),
-            exc_info=True,
-        )
-    log.info(
-        "pipeline_dispatch_complete",
-        game_id=str(game_id),
-        turn_id=str(turn_id),
-        status=result.status,
-        latency_ms=round(elapsed_ms, 1),
-    )
-
-
-async def _generate_title_bg(app_state: object, game_id: UUID, narrative: str) -> None:
-    """Fire-and-forget: generate a title from the opening narrative."""
-    try:
-        svc = app_state.summary_service  # type: ignore[attr-defined]
-        title = await svc.generate_title(narrative)
-        if title:
-            sf = app_state.pipeline_deps.turn_repo._sf  # type: ignore[attr-defined]
-            async with sf() as sess:
-                await sess.execute(
-                    sa.text(
-                        "UPDATE game_sessions SET title = :t, "
-                        "updated_at = :now WHERE id = :gid AND title IS NULL"
-                    ),
-                    {
-                        "gid": game_id,
-                        "t": title[:80],
-                        "now": datetime.now(UTC),
-                    },
-                )
-                await sess.commit()
-            log.info("title_generated", game_id=str(game_id))
-    except Exception:
-        log.warning("title_generation_failed", game_id=str(game_id), exc_info=True)
-
-
-async def _regen_summary_bg(app_state: object, game_id: UUID) -> None:
-    """Fire-and-forget: regenerate the context summary for a game."""
-    try:
-        sf = app_state.pipeline_deps.turn_repo._sf  # type: ignore[attr-defined]
-        turn_repo = app_state.pipeline_deps.turn_repo  # type: ignore[attr-defined]
-        settings = app_state.settings  # type: ignore[attr-defined]
-
-        turns = await turn_repo.get_recent_turns(
-            game_id, limit=settings.resume_turn_count
-        )
-        if not turns:
-            return
-
-        svc = app_state.summary_service  # type: ignore[attr-defined]
-        summary = await svc.generate_context_summary(turns)
-        if summary:
-            now = datetime.now(UTC)
-            async with sf() as sess:
-                await sess.execute(
-                    sa.text(
-                        "UPDATE game_sessions "
-                        "SET summary = :s, summary_generated_at = :now, "
-                        "updated_at = :now WHERE id = :gid"
-                    ),
-                    {"gid": game_id, "s": summary[:200], "now": now},
-                )
-                await sess.commit()
-            log.info("summary_regenerated", game_id=str(game_id))
-    except Exception:
-        log.warning(
-            "summary_regeneration_failed",
-            game_id=str(game_id),
-            exc_info=True,
-        )
-
-
-async def _write_snapshot_bg(
-    app_state: object,
-    game_id: UUID,
-    game_state_dict: dict,
-    turn_number: int,
-) -> None:
-    """Fire-and-forget: persist a GameState snapshot to PostgreSQL (AC-12.04)."""
-    try:
-        state = GameState.model_validate(
-            {**game_state_dict, "session_id": str(game_id), "turn_number": turn_number}
-        )
-        svc = app_state.snapshot_service  # type: ignore[attr-defined]
-        await svc.save_snapshot(game_id, state)
-    except Exception:
-        log.warning("snapshot_write_failed", game_id=str(game_id), exc_info=True)
-
-
+# ── Command router (S01 AC-1.10) ──
 # Valid state transitions (plan §6.1, S27 FR-27.01)
 _VALID_TRANSITIONS: dict[str, set[str]] = {
     "created": {"active", "abandoned"},
@@ -1528,7 +1033,7 @@ async def submit_turn(
     if isinstance(game_state, str):
         game_state = json.loads(game_state)
     asyncio.create_task(
-        _dispatch_pipeline(
+        dispatch_pipeline(
             app_state=request.app.state,
             game_id=game_id,
             turn_id=turn_id,
@@ -1753,7 +1258,7 @@ async def stream_turn(
 
         # S10 §6.2: state_update for world changes follows narrative_end
         if result.world_state_updates:
-            world_changes = _translate_world_updates(result.world_state_updates)
+            world_changes = translate_world_updates(result.world_state_updates)
             if world_changes:
                 await transport.send_state_update(world_changes)
                 for r in _pending:
